% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily2annual.R
\name{daily2annual}
\alias{daily2annual}
\title{Calculate annual data from daily data}
\usage{
daily2annual(dailyData, FUN = mean, startYear = "01",
  threshold = 0.1, ...)
}
\arguments{
\item{dailyData}{[zoo] zoo object with the daily time series. The date format
must be "\%Y-\%m-\%d"}

\item{FUN}{[function] function that have to be applied for the annual
aggregation (by default: mean)}

\item{startYear}{[character] indicate the month to start the hydrological
year. The format must be "MM" (by default, calendar year: "01")}

\item{threshold}{[numeric] threshold of missing values authorized to compute
the annual aggregation. Values must be between 0 and 1 (by default, 10\%: 0.1)}

\item{...}{further arguments}
}
\value{
\emph{\code{annualData}} [zoo] zoo object with the annual computed
time series. The date is in the format "\%Y-MM-01", corresponding of the start
month of the year
}
\description{
Transform a daily time series into an annual one. The user can give a
threshold of missing values authorized per year for the aggregation and choose
the day to start the year (hydrological year or not)
}
\examples{
daily2annual(dailyprec, FUN = sum, startYear = "08", threshold = 0.1)
}
\author{
Florine Garcia (florine.garcia@gmail.com)

Pierre L'Hermite (pierrelhermite@yahoo.fr)
}
