% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Template_function.R
\name{function_name}
\alias{function_name}
\title{Short resume of function (max 10 words)}
\usage{
function_name(data)
}
\arguments{
\item{name}{[class] of the parameter 1 and details}

\item{name}{[class] of the parameter 2 and details}
}
\value{
\emph{output 1} [class] : explanation [unit]

\emph{output 2} [class] : explanation [unit]
}
\description{
Explain the function
}
\details{

}
\examples{
How to use function

}
\references{

}
\seealso{
\code{\link[package name]{function name}}
}
\author{
Florine Garcia (florine.garcia@gmail.com)

Pierre L'Hermite (pierrelhermite@yahoo.fr)
}
