% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanning.R
\name{hanning}
\alias{hanning}
\title{Filter data with Hanning filter order 2}
\usage{
hanning(monthly_data)
}
\arguments{
\item{monthly_data}{[zoo] rainfall daily, monthly, seasonnality or annual data
in zoo class with date in \%Y-\%m-\%d}
}
\value{
\emph{\code{hanzoo}} [zoo] vector with data after hanning filter 
with date in \%Y-\%m-\%d
}
\description{
Eliminate variations to data series to observe better the long term variation
}
\examples{
## Loading data
data("dailyPrec_bv1")

## Index
result <- hanning(dailyPrec)

}
\references{
A.A. Assani, (1999)
Analysis of rainfall variability (1916-1996) at Lubumbashi (Congo-Kinshasa) 
relative to some atmospheric (southern oscillation) and oceanic
(El Nino / La Nina) circulation indicators
Sécheresse, vol. 10, n° 4, p. 245-252.
\url{http://www.jle.com/fr/revues/sec/e-docs/analyse_de_la_variabilite_temporelle_des_precipitations_1916_1996_a_lubumbashi_congo_kinshasa_en_relation_avec_certains_indicateurs_de_la_circulation_atmospherique_oscillation_australe_et_oceanique_el_ni_o_la_ni_a__25631/article.phtml}
}
\seealso{
\code{\link[piflowtest]{plot_trend}} function to plot
}
\author{
Florine Garcia (florine.garcia@gmail.com)

Pierre L'Hermite (pierrelhermite@yahoo.fr)
}
