% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isDate.R
\name{isDate}
\alias{isDate}
\title{Test date class}
\usage{
isDate(test_class)
}
\arguments{
\item{test_class}{[] the class you will test}
}
\value{
\enumerate{
\emph{\code{class}} the logical vector indicating if the test_class is in
date class (boolean) \cr
\item Return \emph{\code{TRUE}} if a test_class is a date class
\item Return \emph{\code{FALSE}} if a test_class is another class
}
}
\description{
Test if a data is a date class
}
\examples{
isDate(2017-01-02)
isDate(as.Date("2017-01-02"))
}
\author{
Florine Garcia (florine.garcia@gmail.com)

Pierre L'Hermite (pierrelhermite@yahoo.fr)
}
