% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isLeapYear.R
\name{isLeapYear}
\alias{isLeapYear}
\title{Test if a year is a leap year}
\usage{
isLeapYear(year)
}
\arguments{
\item{year}{[numeric] the value of the year(s) to be tested}
}
\value{
\emph{\code{leapYear}} [boolean] the logical vector indicating if
the year(s) is(are) a leap year(s)
}
\description{
Return TRUE if a year is a leap year
}
\examples{
isLeapYear(2017)
isLeapYear(c(2004, 2012, 2015))
}
\author{
Florine Garcia (florine.garcia@gmail.com)

Pierre L'Hermite (pierrelhermite@yahoo.fr)
}
