% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issp.R
\name{issp}
\alias{issp}
\title{Calculate Indice Standardisé des Sécheresses Pluviométriques of Pita (ISSP)
Standardized of rainfall drought index in english.}
\usage{
issp(monthly_data)
}
\arguments{
\item{monthly_data}{[zoo] rainfall monthly data in zoo class
with date in \%Y-\%m-\%d}
}
\value{
List that contains
\itemize{
\item \emph{$issp} [zoo] zoo with the issp values with date in \%Y-\%m-\%d
\item \emph{$length_zoo} [zoo] zoo with the length of drought with date in
\%Y-\%m-\%d [day]
\item \emph{$drought_type} [zoo] zoo with the type of the period for each month
\item \emph{$drought_number} [data.frame] dataframe with the number of different
period by type:
\itemize{
\item Extwet (issp > 2)\cr
\item Verywet (1.99 > issp > 1.5)\cr
\item Wet (1.49 > issp > 1)\cr
\item Normal (0.99 > issp > -0.99)\cr
\item Dry (-1 > issp > -1.49)\cr
\item VeryDry (-1.5 > issp > -1.99)\cr
\item ExtDry (-2 > issp)}
}
}
\description{
Calculate ISSP with rainfall anomaly with the mean rainfall by month 
with the length, the drought type and the intensity
}
\examples{
## Loading data
data("dailyPrec_bv1")

## Index
result <- issp(dailyPrec)

## Plot index
plot_trend(result$issp, trend = TRUE, data_kind = "Precipitation",
name = PluvioData$PluvioName, axis_name_x = "Date",
axis_name_y = Monthly precipitation [mm/month], midvalue = 0)

}
\references{
Pita, M.F. (2000) Un nouvel indice de sécheresse pour les domaines
méditerranéens. Application au bassin de Guadalquivir (sud-ouest de l'Espagne).
\emph{Plubications de l'Association Internationale de Climatologie, 13}, 225-234.
\url{https://idus.us.es/xmlui/bitstream/handle/11441/32523/nice_2000_actes.pdf?sequence=1}
}
\seealso{
\code{\link[piflowtest]{plot_trend}} plot the index
}
\author{
Florine Garcia (florine.garcia@gmail.com)

Pierre L'Hermite (pierrelhermite@yahoo.fr)
}
