% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issq.R
\name{issq}
\alias{issq}
\title{Calculate ISSQ index of Pita for discarge}
\usage{
issq(monthly_data)
}
\arguments{
\item{monthly_data}{[zoo] discharge monthly data in zoo class with date 
in \%Y-\%m-\%d}
}
\value{
List that contains
\itemize{
\item \emph{$issq} [zoo] zoo with the issq values with date in \%Y-\%m-\%d
\item \emph{$length_zoo} [zoo] zoo with the length of drought with date in
\%Y-\%m-\%d [day]
\item \emph{$drought_type} [zoo] zoo with the type of the period for each month
\item \emph{$drought_number} [data.frame] dataframe with the number of different
period by type:
\itemize{
\item Extwet (issq > 2)\cr
\item Verywet (1.99 > issq > 1.5)\cr
\item Wet (1.49 > issq > 1)\cr
\item Normal (0.99 > issq > -0.99)\cr
\item Dry (-1 > issq > -1.49)\cr
\item VeryDry (-1.5 > issq > -1.99)\cr
\item ExtDry (-2 > issq)}
}
}
\description{
Calculate ISSQ with discharge anomaly with quantile 50 by month with the
length, the drought type and the intensity
}
\examples{
## Loading data
data("dailyPrec_bv1")

## Index
result <- issq(dailyPrec)

## Plot index
plot_trend(result$issq, trend = TRUE, data_kind = "Flow",
name = PluvioData$PluvioName, axis_name_x = "Date",
axis_name_y = Monthly flow [mm/month], midvalue = 0)

}
\references{

}
\seealso{
\code{\link[piflowtest]{plot_trend}} plot the index
}
\author{
Florine Garcia (florine.garcia@gmail.com)

Pierre L'Hermite (pierrelhermite@yahoo.fr)
}
