% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_sen_pettitt.R
\name{mk_sen_pettitt}
\alias{mk_sen_pettitt}
\title{Calculate Mann-Kendall, Pettitt, Sen tests}
\usage{
mk_sen_pettitt(data)
}
\arguments{
\item{data}{[zoo] vector with daily, monthly, seasonnal or annual 
with date in \%Y-\%m-\%d}
}
\value{
List that contains
\itemize{
\item \emph{$res_mk} [list] resume of MK test
\item \emph{$symbol_mk} [zoo] vector with MK symbol according 
to pvalue of MK test:
\itemize{
\item "-" if pvalue > 0.1\cr
\item "+" if 0.1 > pvalue > 0.05\cr
\item "++" if 0.05 > pvalue > 0.01\cr
\item "+++" if pvalue < 0.01\cr
}
\item \emph{$pettitt} [list] resume of Pettitt test
\item \emph{$symbol_pettitt} [zoo] vector with Pettitt symbol according to
Pettitt test:
\itemize{
\item "-" if pvalue > 0.1\cr
\item "+" if 0.1 > pvalue > 0.05\cr
\item "++" if 0.05 > pvalue > 0.01\cr
\item "+++" if pvalue < 0.01\cr
}
\item\emph{$slope} [numeric] slope according to Sen test
}
}
\description{
Obtain trend with Mann-Kendall test, breaking point with 
Pettitt test and Sen test to have the coefficient of regression
}
\examples{
#'## Loading data
data("dailyPrec_bv1")

result <- mk_sen_pettitt(dailyPrec)
}
\author{
Florine Garcia (florine.garcia@gmail.com)

Pierre L'Hermite (pierrelhermite@yahoo.fr)
}
