% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly2annual.R
\name{monthly2annual}
\alias{monthly2annual}
\title{Calculate annual data from monthly data}
\usage{
monthly2annual(monthlyData, startYear = "01-01", FUN = mean,
  na.rm = TRUE, threshold = 0.1)
}
\arguments{
\item{monthlyData}{[zoo] zoo object with the monthly time series. The date
format must be "\%Y-\%m-\%d"}

\item{startYear}{[character] indicate the month to start the hydrological
year. The format must be "MM-01" (by default, calendar year: "01")}

\item{FUN}{[function] function that have to be applied for the annual
aggregation (by default: mean)}

\item{na.rm}{[boolean] boolean vector (by default, TRUE)}

\item{threshold}{[numeric] threshold of missing values authorized to compute
the annual aggregation. Values must be between 0 and 1 (by default, 10\%: 0.1)}

\item{...}{further arguments}
}
\value{
\emph{\code{annualData}} [zoo] zoo object with the annual computed
time series. The date is in the format "\%Y-MM-01", corresponding of the start
month of the year
}
\description{
Transform a monthly time series into an annual one. The user can give a
threshold of missing values authorized per year for the aggregation and choose
the day to start the year (hydrological year or not)
}
\examples{
monthly2annual(monthlyprec, FUN = sum, startYear = "08-01", threshold = 0.1,
na.rm =TRUE)
}
\author{
Florine Garcia (florine.garcia@gmail.com)

Pierre L'Hermite (pierrelhermite@yahoo.fr)
}
