% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbNoNAs.R
\name{nbNoNAs}
\alias{nbNoNAs}
\title{Calculate the number of dates without missing values per year or month}
\usage{
nbNoNAs(tsData, tstp = "years", hydroYear = NULL)
}
\arguments{
\item{tsData}{[zoo] zoo object with the daily or monthly time series. The
date format must be:
\itemize{  
  \item "\%Y-\%m-\%d" for daily data
  \item "\%Y-\%m-01" for monthly data
}}

\item{tstp}{[character] indicate the period for the calculation. Could be:
\itemize{  
  \item "years": to calculate the number of days or months per year
  \item "months": to calculate the number of days per month
  \item "mPerY": to calculate the number of days per month per year
}}

\item{hydroYear}{[character] vector of the translated years when calculating
the number of dates per hydrological year (by default, calendar year: NULL)}
}
\value{
\emph{missV} [zoo] zoo object with the annual or monthly computed
time series of number of dates without missing values. The date is in the
format: "\%Y-MM-01", corresponding of the start of the year
\itemize{  
  \item "\%Y-MM-01" for annual computation, corresponding of the start month of the year
  \item "\%Y-\%m-01" for monthly computation
}
}
\description{
For a daily or a monthly time series, calculate the number of days without
missing values per year or per month
}
\author{
Florine Garcia (florine.garcia@gmail.com)

Pierre L'Hermite (pierrelhermite@yahoo.fr)
}
