% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnpi.R
\name{pnpi}
\alias{pnpi}
\title{Calculate Percent of Normal Index (PNPI)}
\usage{
pnpi(monthly_data, time_step = 12)
}
\arguments{
\item{monthly_data}{[zoo] rainfall monthly data in zoo class 
with date in \%Y-\%m-\%d}

\item{time_step}{[numeric] default = 12, time step (1, 3, 6, 9, 12, 24 and 48)}
}
\value{
list that contains
\itemize{
\item \emph{pnpi} [zoo] zoo with the pnpi values with date in %Y-\%m-\%d
\item \emph{length_zoo} [zoo] zoo with the length of drought with date in
\%Y-\%m-\%d [day]
\item \emph{drought_type} [zoo] zoo with the type of the period for each month
\item \emph{drought_number} [data.frame] dataframe with the number of
different period by type:
\itemize{
\item Extwet (pnpi >= 160)\cr
\item Verywet (160 > pnpi >= 145)\cr
\item Wet (145 > pnpi >= 120)\cr
\item Normal (120 > pnpi >= 80)\cr
\item Dry (80 > pnpi >= 55)\cr
\item VeryDry (50 > pnpi >= 40)\cr
\item ExtDry (40 > pnpi)}
}
}
\description{
Calculate PNPI and the drought specifications with the length, the drought 
type and the intensity
}
\examples{
## Loading data
data("dailyPrec_bv1")

## Index
result <- pnpi(dailyPrec, time_step = 3)

## Plot index
plot_trend(result$pnpi, trend = TRUE, data_kind = "PNPI",
name = PluvioData$PluvioName, axis_name_x = "Date",
axis_name_y = Monthly precipitation [mm/month], midvalue = 100)

}
\references{
Willeke G, Hosking JRM, Wallis JR, Guttman NB (1994) The national drought 
atlas,\emph{Institute for water resources report 94-NDs-4. Army Corps of 
Engineers}, Washington, D.C.\cr
Amirataee, B. and Montaseri, M. (2017) The performance of SPI and PNPI in analyzing
the spatial and temporal trend of dry and wet periods over Iran, \emph{Natural Hazards},
1, 89-106.
\url{https://doi.org/10.1007/s11069-016-2675-4}
}
\seealso{
\code{\link[piflowtest]{plot_trend}} plot the index
}
\author{
Florine Garcia (florine.garcia@gmail.com)

Pierre L'Hermite (pierrelhermite@yahoo.fr)
}
