% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rai.R
\name{rai}
\alias{rai}
\title{Calculate Rainfall Anomaly Index (RAI)}
\usage{
rai(monthly_data, time_step = 12)
}
\arguments{
\item{monthly_data}{[zoo] rainfall monthly data in zoo class
with date in \%Y-\%m-\%d}

\item{time_step}{[numeric] default = 12, time step to sum monthly 
precipitation (1, 3, 6, 9, 12, 24 and 48)}
}
\value{
List that contains
\itemize{
\item \emph{$rai} [zoo] zoo with the rai values with date in \%Y-\%m-\%d
\item \emph{$length_zoo} [zoo] zoo with the length of drought with date in
\%Y-\%m-\%d [day]
\item \emph{$drought_type} [zoo] zoo with the type of the period for each month
\item \emph{$drought_number} [data.frame] dataframe with the number of different
period by type:
\itemize{
\item Extwet (rai > 3)\cr
\item Verywet (2.99 > rai > 2)\cr
\item Wet (1.99 > rai > 1)\cr
\item Normal (0.99 > rai > -0.99)\cr
\item Dry (-1 > rai > -1.99)\cr
\item VeryDry (-2 > rai > -2.99)\cr
\item ExtDry (-3 > rai)}
}
}
\description{
Calculate RAI and the drought specifications with the length, the drought 
type and the intensity
}
\examples{
## Loading data
data("dailyPrec_bv1")

## Index
result <- rai(dailyPrec, time_step = 12)

## Plot index
plot_trend(result$rai, trend = TRUE, data_kind = "RAI",
name = PluvioData$PluvioName, axis_name_x = "Date",
axis_name_y = RAI [], midvalue = 0)

}
\references{
Van Rooy, M. P. (1965). A rainfall anomaly index independent of time and space.
\emph{Notos, 14}, 43–48.
}
\seealso{
\code{\link[piflowtest]{plot_trend}} plot the index
}
\author{
Florine Garcia (florine.garcia@gmail.com)

Pierre L'Hermite (pierrelhermite@yahoo.fr)
}
