% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spei.R
\name{spei}
\alias{spei}
\title{Calculate Standardized Precipitation Evapotranspiration Index (SPEI)}
\usage{
spei(prec_data, evapo_data, time_step = 12,
  distribution = "log-Logistic")
}
\arguments{
\item{prec_data}{[zoo] rainfall monthly data in zoo class with date 
in \%Y-\%m-\%d}

\item{evapo_data}{[zoo] evapotranspiration monthly data in zoo class 
with date in \%Y-\%m-\%d}

\item{time_step}{[numeric] by default = 12, time step to sum monthly data
(1, 3, 6, 9, 12, 24 and 48)}

\item{distribution}{[character] distribution of data 
(log_Logistic, gamma, grev, genlog, normal)}
}
\value{
list that contains

\emph{spei} [zoo] zoo with the spei values with date in \%Y-\%m-\%d

\emph{drought_type} [zoo] zoo with the type of the period for each
month

\emph{drought_number} [data.frame] dataframe with the number of 
different period by type
\itemize{
\item Extwet (spei > 2)\cr
\item Verywet (1.99 > spei > 1.5)\cr
\item Wet (1.49 > spei > 1)\cr
\item Normal (0.99 > spei > -0.99)\cr
\item Dry (-1 > spei > -1.49)\cr
\item VeryDry (-1.5 > spei > -1.99)\cr
\item ExtDry (-2 > spei)
}
}
\description{
Calculate SPEI and the drought specifications with the length, the drought 
type and the intensity
}
\examples{
How to use function

}
\references{
Vincente-Serrano, S.M. et al, (2010) A multiscalar drought index sensitive 
to global warming: the standardized precipitation evapotranspiration index.
\emph{Journal of Climate, 23}
\url{https://www.researchgate.net/profile/Sergio_Vicente-Serrano/publication/262012840_A_Multiscalar_Drought_Index_Sensitive_to_Global_Warming_The_Standardized_Precipitation_Evapotranspiration_Index/links/540c6b1d0cf2f2b29a377f27/A-Multiscalar-Drought-Index-Sensitive-to-Global-Warming-The-Standardized-Precipitation-Evapotranspiration-Index.pdf}
}
\seealso{
\code{\link[piflowtest]{plot_trend}}: plot the index
}
\author{
Florine Garcia (florine.garcia@gmail.com)

Pierre L'Hermite (pierrelhermite@yahoo.fr)
}
