% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi.R
\name{spi}
\alias{spi}
\title{Calculate Standardized Precipitation Index (SPI)}
\usage{
spi(prec_data, time_step = 12, distribution = "gamma")
}
\arguments{
\item{prec_data}{[zoo] rainfall monthly data in zoo class with date 
in \%Y-\%m-\%d}

\item{time_step}{[numeric] by default = 12, time step to sum monthly data
(1, 3, 6, 9, 12, 24 and 48)}

\item{distribution}{[character] distribution of data 
(log_Logistic, gamma, grev, genlog, normal)}
}
\value{
list that contains

\emph{spi} [zoo] zoo with the spi values with date in \%Y-\%m-\%d

\emph{drought_type} [zoo] zoo with the type of the period for each
month

\emph{drought_number} [data.frame] dataframe with the number of 
different period by type
\itemize{
\item Extwet (spi > 2)\cr
\item Verywet (1.99 > spi > 1.5)\cr
\item Wet (1.49 > spi > 1)\cr
\item Normal (0.99 > spi > -0.99)\cr
\item Dry (-1 > spi > -1.49)\cr
\item VeryDry (-1.5 > spi > -1.99)\cr
\item ExtDry (-2 > spi)
}
}
\description{
Calculate SPI and the drought specifications with the length, the drought 
type and the intensity
}
\examples{
How to use function

}
\references{
McKee, T.B., Doesken, N.J., & Kleist, J. (1993)
The relationship of drought frequency and duration of time scales.
\emph{Eighth Conference on Applied Climatology, American Meteorological Society}
\url{http://www.droughtmanagement.info/literature/AMS_Relationship_Drought_Frequency_Duration_Time_Scales_1993.pdf}
}
\seealso{
\code{\link[piflowtest]{plot_trend}}: plot the index
}
\author{
Florine Garcia (florine.garcia@gmail.com)

Pierre L'Hermite (pierrelhermite@yahoo.fr)
}
