% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssfi.R
\name{ssfi}
\alias{ssfi}
\title{Calculate Standardized StreamFlow Index (SSFI)}
\usage{
ssfi(prec_data, Delta = 12, Distribution = "gamma")
}
\arguments{
\item{prec_data}{[zoo] rainfall monthly data in zoo class with date 
in \%Y-\%m-\%d}

\item{time_step}{[numeric] by default = 12, time step to sum monthly data
(1, 3, 6, 9, 12, 24 and 48)}

\item{distribution}{[character] distribution of data 
(log_Logistic, gamma, grev, genlog, normal)}
}
\value{
list that contains

\emph{ssfi} [zoo] zoo with the ssfi values with date in \%Y-\%m-\%d

\emph{drought_type} [zoo] zoo with the type of the period for each
month

\emph{drought_number} [data.frame] dataframe with the number of 
different period by type
\itemize{
\item Extwet (ssfi > 2)\cr
\item Verywet (1.99 > ssfi > 1.5)\cr
\item Wet (1.49 > ssfi > 1)\cr
\item Normal (0.99 > ssfi > -0.99)\cr
\item Dry (-1 > ssfi > -1.49)\cr
\item VeryDry (-1.5 > ssfi > -1.99)\cr
\item ExtDry (-2 > ssfi)
}
}
\description{
Calculate SSFI and the drought specifications with the length, the drought 
type and the intensity
}
\examples{
How to use function

}
\references{
Modarres, R., (2007) Streamflow drought time series forecasting. 
\emph{Stochastic Environmental Research and Risk Assessment}
21: 223–233.
\url{https://doi.org/10.1007/s00477-006-0058-1}
}
\seealso{
\code{\link[piflowtest]{plot_trend}}: plot the index
}
\author{
Florine Garcia (florine.garcia@gmail.com)

Pierre L'Hermite (pierrelhermite@yahoo.fr)
}
