% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsExtension.R
\name{tsExtension}
\alias{tsExtension}
\title{Extend a daily or monthly time series}
\usage{
tsExtension(tsData, startDate = start(tsData), endDate = end(tsData))
}
\arguments{
\item{tsData}{[zoo] zoo object with the daily or monthly time series.
The date format must be:
\itemize{
 \item{"\%Y-\%m-\%d" for daily data}
 \item{"\%Y-\%m-01" for monthly data}
}}

\item{startDate}{[Date] the first date of the extension. The date format 
must be:
\itemize{
 \item{"\%Y-\%m-\%d" for daily data}
 \item{"\%Y-\%m-01" for monthly data}
} (by default, first date of tsData: start(tsData))}

\item{endDate}{[Date] the last date of the extension. The date format 
must be:
\itemize{
 \item{"\%Y-\%m-\%d" for daily data}
 \item{"\%Y-\%m-01" for monthly data}
} (by default, last date of tsData: end(tsData))}
}
\value{
extTsData [zoo] regular zoo object with the daily or monthly extended 
time series. The date format is:
\itemize{
 \item{"\%Y-\%m-\%d" for daily data}
 \item{"\%Y-\%m-01" for monthly data}
}
}
\description{
Extension of a daily or monthly time series with the chosen first and 
last dates, and transformation of irregular zoo in regular zoo by 
adding NA when missing dates.
}
\author{
Florine Garcia (florine.garcia@gmail.com)

Pierre L'Hermite (pierrelhermite@yahoo.fr)
}
