% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearTranslation.R
\name{yearTranslation}
\alias{yearTranslation}
\title{Create a vector for the computation of hydrological years}
\usage{
yearTranslation(tsData, startYear)
}
\arguments{
\item{tsData}{[zoo] zoo object with the daily or monthly time series.
The date format must be:
\itemize{
 \item{"\%Y-\%m-\%d" for daily data}
 \item{"\%Y-\%m-01" for monthly data}
}}

\item{startYear}{[character] indicate the month to start the hydrological year.
The format must be "MM" (by default, calendar year: "01")}
}
\value{
years [character] vector of the translated years
}
\description{
Creation of a vector years translated to compute statistics on
hydrological years
}
\author{
Florine Garcia (florine.garcia@gmail.com)

Pierre L'Hermite (pierrelhermite@yahoo.fr)
}
